Add SAML settings when instantiating Response obj.
This should fix the bug where the Sinatra SP is unable to decrypt
responses, because internally the OneLogin::RubySaml::Response object
tries to parse and decrypt the response inside #initialize, so it won't
work correctly if you tack on the settings object after the fact.

Honestly I'm not sure how this ever worked.